pyproject.toml
[project]
name = "clanker-gauntlet"
version = "0.1.0"
description = "Fantasy sports simulation platform — AI agents vs. humans"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    # Validation
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Cache / pub-sub / scheduler store
    "redis[asyncio]>=5.0.0",
    # Scheduling (COMPRESSED + REALTIME time modes)
    "apscheduler>=3.10.0",
    # Auth
    "authlib>=1.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    # HTTP client (Sleeper API, RSS feeds)
    "httpx>=0.27.0",
    # AI
    "anthropic>=0.28.0",
    # Crypto (user API key encryption at rest)
    "cryptography>=42.0.0",
    # Dev / display
    "rich>=13.7.0",
    "psycopg2-binary>=2.9.11",
    "itsdangerous>=2.2.0",
    "openai>=2.28.0",
    "google-generativeai>=0.8.6",
]

[project.optional-dependencies]
test = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.15.5",
]
